tests: Fix a format error on darwin
test-compositor.c:94:25: error: format specifies type 'long' but the argument has type '__darwin_suseconds_t' (aka 'int') [-Werror,-Wformat]
                 getpid(), tv.tv_sec, tv.tv_usec);
                                      ^~~~~~~~~~

+++ tests/test-compositor.c
@@ -91,7 +91,7 @@ get_socket_name(void)
 
 	gettimeofday(&tv, NULL);
 	snprintf(retval, sizeof retval, "wayland-test-%d-%ld%ld",
-		 getpid(), tv.tv_sec, tv.tv_usec);
+		 (int)getpid(), (long)tv.tv_sec, (long)tv.tv_usec);
 
 	return retval;
 }
